python: Refactor params.py into a package
params.py was getting quite unwieldy. This change refactors it into multiple
files in the params package. There should be no user-facing changes due
to this change. Subsequent commits will do some cleanup. This commit should
only be moving code around.
