+++ src/python/SConscript
@@ -369,12 +369,21 @@ PySource('m5', 'm5/debug.py')
 PySource('m5', 'm5/event.py')
 PySource('m5', 'm5/main.py')
 PySource('m5', 'm5/options.py')
-PySource('m5', 'm5/params.py')
 PySource('m5', 'm5/proxy.py')
 PySource('m5', 'm5/simulate.py')
 PySource('m5', 'm5/ticks.py')
 PySource('m5', 'm5/trace.py')
 PySource('m5.objects', 'm5/objects/__init__.py')
+PySource('m5.params', 'm5/params/__init__.py')
+PySource('m5.params', 'm5/params/base_params.py')
+PySource('m5.params', 'm5/params/deprecated_params.py')
+PySource('m5.params', 'm5/params/enum_params.py')
+PySource('m5.params', 'm5/params/network_params.py')
+PySource('m5.params', 'm5/params/null_params.py')
+PySource('m5.params', 'm5/params/param_types.py')
+PySource('m5.params', 'm5/params/port_params.py')
+PySource('m5.params', 'm5/params/power_params.py')
+PySource('m5.params', 'm5/params/time_params.py')
 PySource('m5.stats', 'm5/stats/__init__.py')
 PySource('m5.util', 'm5/util/__init__.py')
 PySource('m5.util', 'm5/util/attrdict.py')
